Example data about mental health descirptions .

Example data about mental health descirptions .

Usage

dep_wor_data

Format

A data frame with 500 participants and 13 variables:

Depselect

Words that respondents have selected from a pre-defined list

Worselect

Words that respondents have selected from a pre-defined list

Depword

Wrods where respondents describe their experience with depression in life

Worword

Words where respondents describe their experience with depression in life

Depphrase

phrases where respondents describe their experience with depression in life

Worphrase

Phrases where respondents describe their experience with anxiety in life

Deptext

Text where respondents describe their experience with depression in life

Wortext

Text where respondents describe their experience with anxiety in life

Gender

Respondents gender 0=male, 1=female

Age

respondents age in years

PHQ9tot

total score of the respondents PHQ-9

GAD7tot

total score of the respondents GAD-7

MentalSickDays

self-reported sick days due to mental health issues

MentalSickDaysYear

self-reported sick days due to mental health issues in the last year

HealthCareMental

self-reported health care visits due to mental health

SickDaysMonth

self-reported sick days last month

SickDaysYear

self-reported sick days last year

HealthCareVisits

self-reported health care visits

Source

<https://osf.io/preprints/psyarxiv/p67db>