The function to create lda wordclouds

The function to create lda wordclouds

topicsPlot(
  model = NULL,
  ngrams = NULL,
  test = NULL,
  p_threshold = 0.05,
  grid_plot = TRUE,
  dim = 2,
  color_scheme = "default",
  scatter_legend_popout_num = c(1, 1, 1, 1, 0, 1, 1, 1, 1),
  scatter_legend_way_popout_topics = c("mean", "max_x", "max_y"),
  scatter_legend_user_spec_topics = NULL,
  scatter_legend_topic_num = FALSE,
  scale_size = FALSE,
  plot_topics_idx = NULL,
  save_dir = "./results",
  figure_format = "svg",
  width = 10,
  height = 8,
  max_size = 10,
  seed = 42,
  scatter_legend_popout_dot_size = 15,
  scatter_legend_bg_dot_size = 9,
  grid_legend_title = "legend_title",
  grid_legend_title_size = 5,
  grid_legend_title_color = "black",
  grid_legend_x_axes_label = "legend_x_axes_label",
  grid_legend_y_axes_label = "legend_y_axes_label",
  grid_legend_number_color = "black",
  grid_legend_number_size = 5
)

Arguments

model

(list) The trained model

test

(list) The test results

p_threshold

(integer) The p-value threshold to use for significance

grid_plot

(boolean) Set TRUE if plotting the topics grid

dim

(numeric) Generate 1 dimensional color plots or 2 dimensional color plots if grid = TRUE

color_scheme

(string 'default' or vector) The color scheme for plotted topic categories if grid = TRUE. The vector should contain 9 color codes.

scatter_legend_popout_num

(numeric or vector) The vector of num of pop outs of each grid in the scatter legend.

scatter_legend_way_popout_topics

(string) The way to filter topics to pop out in the scatter legend. Can be either "mean", "max_x", or "max_y"

scatter_legend_user_spec_topics

(vector) User can specify which topic to be poped out in the scatter legend. Should be like c("t_1", "t_2", ...). If set, way_popout_topics will have no effect.

scatter_legend_topic_num

(boolean) Allow showing the topic number or not in the scatter legend

scale_size

(logical) Whether to scale the size of the words

save_dir

(string) The directory to save the wordclouds

figure_format

(string) Set the figure format, e.g., .svg, or .png.

width

(integer) The width of the topic (units = "in").

height

(integer) The width of the topic (units = "in").

max_size

(integer) The max size of the words.

seed

(integer) The seed to set for reproducibility

scatter_legend_popout_dot_size

(integer) The dot size of scatter legend to popout

scatter_legend_bg_dot_size

(integer) The dot size of scatter legend in the background

grid_legend_title

The title of grid topic plot if grid_plot = TRUE

grid_legend_title_size

The size of the title of the plot if grid_plot = TRUE

grid_legend_title_color

The color of the legend title if grid_plot = TRUE

grid_legend_x_axes_label

The label of the x axes if grid_plot = TRUE

grid_legend_y_axes_label

The label of the y axes if grid_plot = TRUE

grid_legend_number_color

The color in the text in the legend if grid_plot = TRUE

grid_legend_number_size

The color in the text in the legend if grid_plot = TRUE

Value

nothing is returned, the wordclouds are saved in the save_dir